Module Patience_diff_lib.Patience_diff

module Hunk : sig ... end

In diff terms, a hunk is a unit of consecutive ranges with some Same context before and after Next, Prev, and Replace ranges. Each hunk contains information about the original arrays, specifically the starting indexes and the number of elements in both arrays to which the hunk refers.

module Hunks : sig ... end

module Matching_block : sig ... end

module Range : sig ... end

For handling diffs abstractly. A range is a subarray of the two original arrays with a constructor defining its relationship to the two original arrays. A Same range contains a series of elements which can be found in both arrays. A Next range contains elements found only in the second array, while an Prev range contains elements found only in the first array.

module Make (Elt : Core.Hashtbl.Key) : sig ... end

module String : sig ... end

module Stable : sig ... end